collect-data/requirement.txt
kafka-python
hdfs
pandas
